Rewrote Config classes and deprecated old ones.
We now have a ConfigBag (with interface) that is used for holding and
accessing config values. ConfigBag uses Bags from "bolt/collection"
library which handles the nested path-like syntax for us.

We have a ConfigFactory which reads and parses files and creates ConfigBags.

And finally, Config singleton, which closely matches the old classes by
creating the ConfigBag and accessing it statically. It also doesn't break
if there are multiple subclasses being used at the same time (like the
old classes did).