Created dependency resolving code to recursively find item dependencies of each other from a list/map.
I'm not thrilled with how the code turned out. There seemed to be some common logic with this process
that didn't need to be implemented for each concrete use case. Such as the recursion, checking for
cyclic dependencies, sorting items for least dependent first, checking that the dependency is in the list.
Ultimately it turned out to be a lot of code to try to handle the differences between lists and maps.